refactor: 코드에서 쓰인 모든 경로 상수화
