+++ src/constants/routes.ts
@@ -1,0 +1,9 @@
+const ROUTES = {
+  HOME: '/',
+  LOGIN: '/login',
+  MY_PAGE: '/my',
+  ORDER_DETAIL_BASE: '/order/:id',
+  NOT_FOUND: '*',
+} as const;
+
+export default ROUTES;
